+++ contracts/common/components/IDocumentsRepo.sol
@@ -44,7 +44,7 @@ interface IDocumentsRepo {
 
     function isRegistered(address body) external view returns (bool);
 
-    function counterOfDocs() external view returns (uint16);
+    function counterOfDocs() external view returns (uint32);
 
     function passedReview(address body) external view returns (bool);
